call fabric types moved to a new file
